refactor(mbk): move test_utils.py out of app/api to app/test_helpers/
Split the 1069-LOC test_utils.py into three thematic modules under a new
app/test_helpers/ package so test-only endpoint code no longer lives
alongside production routes in app/api/:

- app/test_helpers/auth.py   — _require_test_mode guard + promote-admin
- app/test_helpers/seed.py   — all seed/delete E2E data management endpoints
- app/test_helpers/mocks.py  — Gmail send stub, storage mock, reauth state
- app/test_helpers/router.py — unified router that aggregates the three

All lazy imports promoted to module-level in the new files.
main.py now imports from app.test_helpers.router; the conditional env-gated
mount logic is unchanged. Test file updated to import from the new locations
and to patch app.test_helpers.auth.unit_of_work in addition to seed.
